manage inventory condition added- user should have more than 15 products to delete a item

+++ src/Components/ManageInventory/ManageInventory.js
@@ -7,20 +7,17 @@ const ManageInventory = () => {
     const [products, setProducts] = UseItems();
     const handleDelete = id => {
         const proceed = window.confirm('Deleted items can not be recoverd. Do you want to proceed?');
-        if (proceed) {
+        if (proceed && products.length > 15) {
             const url = `https://young-gorge-41222.herokuapp.com/items/${id}`
             fetch(url, {
                 method: 'DELETE'
             }).then(res => res.json()).then(data => {
                 console.log(data);
                 const remaining = products.filter(product => product._id !== id);
-
-                if (remaining.length >= 15) {
-                    setProducts(remaining);
-                } else {
-                    window.confirm('Company have enough land to supply fruits. For This reason we should have at least 15 stock fruits info')
-                }
+                setProducts(remaining);
             });
+        }else{
+            alert('Inventory must have atleast 15 items. Add new items to delete')
         }
     }
     return (
